Fix tmux session names starting with '-'
`bin/tat` command replaces '.' with '-' to make a legal TMUX session
name. But that means if the directory started with '.' (meaning it's a
hidden directory). The leading '-' was causing a problem with toilet
making it miss fonts (and other errors) because it interprets the input
as an option (rather than text to be output).